Add README
This commit adds a README file to the repository with some basic
instructions about what this project is and some tips to get started.
Since we need a binary DLL from the FoundationDB site, I've added
a PowerShell script that can download this and added the instructions
to the README.

Also updated the .gitignore file to exclude the binaries in the native
folder.