Add SafeHandles for FoundationDB handles
FoundationDB exposes several types of handles which need to be cleaned
up reliably. By wrapping these handles in a SafeHandle class the .NET
runtime will help prevent leaking handles.

The handles provided by FoundationDB are:
  1. Futures
  2. Transactions
  3. Databases
  4. Clusters

Almost everything in the FoundationDB API will return a Future object
that represents an asynchronous operation (analogous to a Task in .NET).
These Futures can represent significant amounts of memory that are
allocated as the result of the operation, and freeing that memory
reliably is crucial. Since a lot of operations in the FoundationDB API
are async this allows significant scalability benefits. These Futures
will be wrapped and represented as an awaitable operation (most likely
as a Task).

Transactions are used to correlate operations performed on the database
and are generally interacted with by end users the most. Transaction
handles are obtained from Databases in a synchronous fashion.

Databases are not used often, in the current version of the FoundationDB
ABI (510) only a single database is supported. So generally an end user
will be holding on to only 1 of these.

Clusters represent a set of machines that host FoundationDB databases.
Currently I expect that end users will not be connecting to multiple
FoundationDB clusters simultaneously.