Add infrastructure to handle FdbError codes
Since the correct way to indicate a failure of a method is to throw an
exception instead of returning an error code, we need some
infrastructure to support the conversion of FdbError codes (int) to
exceptions.

A small subset of error codes has been added to the FdbError
enumeration, more work is required to add all the possible error codes
which are listed at [1].

While the enum description of the FdbError enumeration should provide
enough information, the native API also has support for retrieving an
English string. For a good debug experience this error string is
retrieved and cached from the native API and included in the exception.

[1]: https://apple.github.io/foundationdb/api-error-codes.html#developer-guide-error-codes