Add alternative for BITS to Retrieve-NativeFdb.ps1
The Background Intelligent Transfer Service (BITS) is not always
available. On build servers this service might not be started, or
because of the way agents execute it might not be available. When the
BITS service is not running, it will use the .NET WebClient to download
the native libraries.

Also added a check to avoid creating the directory multiple times if it
already exists.