Add network loop holding class
The Fdb class encapsulates the static requirements laid out by the
native FoundationDB library. It requires setting an API version as well
as a so called 'network thread' which allows FoundationDB Futures to
execute. These calls can only be done once for the lifetime of the
application, so to express this the Fdb class is implemented as a
Singleton.

It is advised that consumers of this library call the Fdb.Terminate()
method when they know that their runtime is shutting down. For example:
ASP.NET Core provides the IApplicationLifetime interface that expresses
the ApplicationStopping and ApplicationStopped 'events'. This will allow
the client to gracefully stop.

The library will hook into the AppDomain.DomainUnload and
AppDomain.ProcessExit events to try and perform graceful shutdown
automatically. However since the time that can be spent in these methods
is limited, this is not as reliable as I'd like.

Custom network options have not been implemented yet.