Add OpenCluster functionality
The FoundationDB driver uses a cluster which has one or more databases
as an abstraction of a connection.

Note that it is not actually connecting yet, but that the connection
will be opened during the first read/write in a transaction.

Also note that the current FoundationDB implementation (5.1.7 / 5.2.5 at
time of writing) only supports a single database at this time.